feat(user): autenticação somente usuario logado pode alterar ou deletar ele mesmo

+++ src/app/api/user/[id]/route.ts
@@ -2,9 +2,30 @@ import { prisma } from '@/lib/prisma';
 import { hash } from 'bcrypt';
 import { NextRequest } from 'next/server';
 import { getIdFromRequest } from '@/utils/url';
+import jwt from 'jsonwebtoken';
+
+function getUserFromRequest(request: NextRequest) {
+  const authHeader = request.headers.get('authorization');
+  if (!authHeader) return null;
+  const token = authHeader.split(' ')[1];
+  try {
+    return jwt.verify(token, process.env.JWT_SECRET as string) as {
+      id: string;
+    };
+  } catch {
+    return null;
+  }
+}
 
 export async function GET(request: NextRequest) {
   const id = getIdFromRequest(request);
+
+  const userAuth = getUserFromRequest(request);
+
+  if (!userAuth || userAuth.id !== id) {
+    return new Response('Acesso negado', { status: 403 });
+  }
+
   const user = await prisma.user.findUnique({
     where: { id },
     include: {
@@ -23,18 +44,15 @@ export async function GET(request: NextRequest) {
 
 export async function PUT(request: NextRequest) {
   const id = getIdFromRequest(request);
+
+  const userAuth = getUserFromRequest(request);
+  if (!userAuth || userAuth.id !== id) {
+    return new Response('Acesso negado', { status: 403 });
+  }
+
   const body = await request.json();
 
-  const {
-    name,
-    email,
-    password,
-    avatar,
-    linkedin,
-    github,
-    bio,
-    skills, // Array de skillIds
-  } = body;
+  const { name, email, password, avatar, linkedin, github, bio, skills } = body;
 
   const hashedPassword = password ? await hash(password, 10) : undefined;
 
@@ -50,7 +68,7 @@ export async function PUT(request: NextRequest) {
       bio,
       skills: skills
         ? {
-            deleteMany: {}, // Remove todas as skills atuais
+            deleteMany: {},
             create: skills.map((skillId: string) => ({
               skill: { connect: { id: skillId } },
             })),
@@ -69,6 +87,13 @@ export async function PUT(request: NextRequest) {
 
 export async function DELETE(request: NextRequest) {
   const id = getIdFromRequest(request);
+
+  const userAuth = getUserFromRequest(request);
+
+  if (!userAuth || userAuth.id !== id) {
+    return new Response('Acesso negado', { status: 403 });
+  }
+
   await prisma.user.delete({
     where: { id },
   });
